backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
pytest-django==4.8.0
djangorestframework-simplejwt==5.3.1
django-filter==24.1
drf-spectacular==0.27.1
